fix(db): move ytlive seed to startup + split db/mod.rs and playback/tests.rs
Move the ytlive INSERT from MIGRATION_V13 to a new idempotent
startup::ensure_live_playlist_exists() function. The seed in migrations
caused 16 pre-existing tests to regress (playlist_id=1 collisions,
wrong row counts). Wire the function into lib.rs::start() and update all
tests that relied on the seed being auto-applied.

Extract db/mod.rs inline tests to db/mod_tests.rs and split the two
handle_play_video tests from playback/tests.rs into a new sibling file
playback/tests_play_video.rs, bringing all four files under the 1000-line
airuleset cap.
