+++ crates/sp-server/src/db/mod.rs
@@ -179,10 +179,11 @@ ALTER TABLE videos ADD COLUMN lyrics_manual_priority INTEGER NOT NULL DEFAULT 0;
 // - `playlist_items` stores ordered references to existing videos. Videos
 //   themselves still live under their *home* youtube playlist; this table
 //   just names positions.
-// - The 'ytlive' row is the single pre-created custom playlist used for
-//   tonight's live event. youtube_url is the empty-string sentinel
-//   (the column is NOT NULL; avoiding a table recreate keeps the
-//   migration safe).
+// - The 'ytlive' seed row is intentionally NOT inserted here. Seeding it
+//   in the migration caused pre-existing tests (which call run_migrations on
+//   an empty pool and count rows or hard-code playlist_id=1) to regress.
+//   The row is created by `startup::ensure_live_playlist_exists` instead,
+//   which runs whenever the server actually starts.
 const MIGRATION_V13: &str = "
 ALTER TABLE playlists ADD COLUMN kind TEXT NOT NULL DEFAULT 'youtube';
 ALTER TABLE playlists ADD COLUMN current_position INTEGER NOT NULL DEFAULT 0;
@@ -198,11 +199,6 @@ CREATE TABLE playlist_items (
 );
 CREATE UNIQUE INDEX idx_playlist_items_playlist_video
     ON playlist_items (playlist_id, video_id);
-
-INSERT OR IGNORE INTO playlists
-    (name, youtube_url, ndi_output_name, playback_mode, is_active, kind)
-VALUES
-    ('ytlive', '', 'SP-live', 'continuous', 1, 'custom');
 ";
 
 /// Create a connection pool backed by a file.
@@ -274,809 +270,6 @@ pub async fn current_schema_version(pool: &SqlitePool) -> Result<i32, sqlx::Erro
     Ok(row.get("v"))
 }
 
+#[path = "mod_tests.rs"]
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    async fn setup() -> SqlitePool {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-        pool
-    }
-
-    #[tokio::test]
-    async fn pool_creation_and_migration() {
-        let pool = setup().await;
-        let ver = current_schema_version(&pool).await.unwrap();
-        assert_eq!(ver, 13);
-    }
-
-    #[tokio::test]
-    async fn migrations_are_idempotent() {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-        run_migrations(&pool).await.unwrap(); // second run must not fail
-        let ver = current_schema_version(&pool).await.unwrap();
-        assert_eq!(ver, 13);
-    }
-
-    #[tokio::test]
-    async fn migration_v4_adds_audio_file_path_column() {
-        let pool = setup().await;
-        let cols: Vec<String> = sqlx::query("PRAGMA table_info(videos)")
-            .fetch_all(&pool)
-            .await
-            .unwrap()
-            .iter()
-            .map(|r| r.get::<String, _>("name"))
-            .collect();
-        assert!(
-            cols.contains(&"audio_file_path".to_string()),
-            "audio_file_path column should exist, columns: {cols:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn migration_v4_resets_all_normalized_rows() {
-        let pool = create_memory_pool().await.unwrap();
-        // Apply V1 + V2 + V3 manually so we can seed data before V4.
-        for &(version, sql) in &MIGRATIONS[..3] {
-            // Ensure schema_version table exists so the INSERT below works.
-            sqlx::query(
-                "CREATE TABLE IF NOT EXISTS schema_version (
-                    version INTEGER PRIMARY KEY,
-                    applied_at TEXT NOT NULL DEFAULT (datetime('now'))
-                )",
-            )
-            .execute(&pool)
-            .await
-            .unwrap();
-
-            let mut tx = pool.begin().await.unwrap();
-            for stmt in sql.split(';') {
-                let s = stmt.trim();
-                if !s.is_empty() {
-                    sqlx::query(s).execute(&mut *tx).await.unwrap();
-                }
-            }
-            sqlx::query("INSERT INTO schema_version (version) VALUES (?)")
-                .bind(version)
-                .execute(&mut *tx)
-                .await
-                .unwrap();
-            tx.commit().await.unwrap();
-        }
-        // Seed a playlist and a normalized video.
-        sqlx::query(
-            "INSERT INTO playlists (name, youtube_url, ndi_output_name) VALUES ('p', 'u', 'n')",
-        )
-        .execute(&pool)
-        .await
-        .unwrap();
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id, normalized, file_path) VALUES (1, 'abc', 1, '/tmp/foo.mp4')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Now apply V4.
-        run_migrations(&pool).await.unwrap();
-
-        // Row's normalized must have been reset to 0.
-        let n: i64 = sqlx::query("SELECT normalized FROM videos WHERE id = 1")
-            .fetch_one(&pool)
-            .await
-            .unwrap()
-            .get("normalized");
-        assert_eq!(n, 0, "V4 must reset normalized=0 for every existing row");
-    }
-
-    #[tokio::test]
-    async fn migration_v3_drops_per_playlist_title_columns() {
-        let pool = setup().await;
-        let cols: Vec<String> = sqlx::query("PRAGMA table_info(playlists)")
-            .fetch_all(&pool)
-            .await
-            .unwrap()
-            .iter()
-            .map(|r| r.get::<String, _>("name"))
-            .collect();
-        assert!(
-            !cols.contains(&"obs_text_source".to_string()),
-            "obs_text_source column should be dropped, columns: {cols:?}"
-        );
-        assert!(
-            !cols.contains(&"resolume_title_token".to_string()),
-            "resolume_title_token column should be dropped, columns: {cols:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn migration_creates_all_tables() {
-        let pool = setup().await;
-        let tables: Vec<String> =
-            sqlx::query("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name")
-                .fetch_all(&pool)
-                .await
-                .unwrap()
-                .iter()
-                .map(|r| r.get::<String, _>("name"))
-                .collect();
-
-        for expected in &[
-            "playlists",
-            "videos",
-            "play_history",
-            "settings",
-            "resolume_hosts",
-            "resolume_clip_mappings",
-        ] {
-            assert!(
-                tables.contains(&expected.to_string()),
-                "missing table: {expected}"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn crud_playlist_and_videos() {
-        let pool = setup().await;
-
-        // Insert playlist
-        let res =
-            sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES (?, ?) RETURNING id")
-                .bind("Test Playlist")
-                .bind("https://youtube.com/playlist?list=PLtest")
-                .fetch_one(&pool)
-                .await
-                .unwrap();
-        let playlist_id: i64 = res.get("id");
-        assert!(playlist_id > 0);
-
-        // Insert video
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id, title) VALUES (?, ?, ?)")
-            .bind(playlist_id)
-            .bind("dQw4w9WgXcQ")
-            .bind("Never Gonna Give You Up")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Read back
-        let row = sqlx::query("SELECT title FROM videos WHERE playlist_id = ?")
-            .bind(playlist_id)
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        let title: String = row.get("title");
-        assert_eq!(title, "Never Gonna Give You Up");
-    }
-
-    #[tokio::test]
-    async fn unique_index_enforcement() {
-        let pool = setup().await;
-
-        sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES ('P', 'url')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id) VALUES (1, 'abc')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Duplicate should fail
-        let res = sqlx::query("INSERT INTO videos (playlist_id, youtube_id) VALUES (1, 'abc')")
-            .execute(&pool)
-            .await;
-        assert!(res.is_err(), "duplicate should violate unique index");
-    }
-
-    #[tokio::test]
-    async fn foreign_key_cascade_delete() {
-        let pool = setup().await;
-
-        sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES ('P', 'url')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id) VALUES (1, 'vid1')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        sqlx::query("INSERT INTO play_history (playlist_id, video_id) VALUES (1, 1)")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Delete playlist — should cascade to videos and play_history
-        sqlx::query("DELETE FROM playlists WHERE id = 1")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        let video_count: i64 = sqlx::query("SELECT COUNT(*) AS c FROM videos")
-            .fetch_one(&pool)
-            .await
-            .unwrap()
-            .get("c");
-        assert_eq!(video_count, 0);
-
-        let history_count: i64 = sqlx::query("SELECT COUNT(*) AS c FROM play_history")
-            .fetch_one(&pool)
-            .await
-            .unwrap()
-            .get("c");
-        assert_eq!(history_count, 0);
-    }
-
-    #[tokio::test]
-    async fn settings_roundtrip() {
-        let pool = setup().await;
-
-        models::set_setting(&pool, "obs_url", "ws://127.0.0.1:4455")
-            .await
-            .unwrap();
-
-        let val = models::get_setting(&pool, "obs_url").await.unwrap();
-        assert_eq!(val, Some("ws://127.0.0.1:4455".to_string()));
-
-        // Update
-        models::set_setting(&pool, "obs_url", "ws://10.0.0.1:4455")
-            .await
-            .unwrap();
-        let val = models::get_setting(&pool, "obs_url").await.unwrap();
-        assert_eq!(val, Some("ws://10.0.0.1:4455".to_string()));
-
-        // Non-existent key
-        let missing = models::get_setting(&pool, "nope").await.unwrap();
-        assert_eq!(missing, None);
-    }
-
-    #[tokio::test]
-    async fn model_query_roundtrips() {
-        let pool = setup().await;
-
-        // Insert a playlist via models
-        let playlist = models::insert_playlist(&pool, "Worship", "https://yt.com/pl1")
-            .await
-            .unwrap();
-        assert_eq!(playlist.name, "Worship");
-        assert_eq!(playlist.youtube_url, "https://yt.com/pl1");
-        assert!(playlist.is_active);
-
-        // Get active playlists — verify all fields are populated
-        let active = models::get_active_playlists(&pool).await.unwrap();
-        // V13 seeds ytlive, so we now have 2 active playlists
-        assert_eq!(active.len(), 2);
-        let worship = active
-            .iter()
-            .find(|p| p.name == "Worship")
-            .expect("should find Worship playlist");
-        assert_eq!(worship.id, playlist.id);
-        assert_eq!(worship.name, "Worship");
-        assert_eq!(worship.youtube_url, "https://yt.com/pl1");
-        assert!(worship.is_active);
-
-        // Upsert video
-        let video = models::upsert_video(&pool, playlist.id, "abc123", Some("My Song"))
-            .await
-            .unwrap();
-        assert_eq!(video.youtube_id, "abc123");
-        assert_eq!(video.title, "My Song");
-
-        // Upsert same video with updated title
-        let video2 = models::upsert_video(&pool, playlist.id, "abc123", Some("Updated"))
-            .await
-            .unwrap();
-        assert_eq!(video2.id, video.id); // same row
-        assert_eq!(video2.title, "Updated");
-
-        // Get videos for playlist
-        let videos = models::get_videos_for_playlist(&pool, playlist.id)
-            .await
-            .unwrap();
-        assert_eq!(videos.len(), 1);
-
-        // Record play
-        models::record_play(&pool, playlist.id, video.id)
-            .await
-            .unwrap();
-
-        // Get unplayed — should be empty now
-        let unplayed = models::get_unplayed_video_ids(&pool, playlist.id)
-            .await
-            .unwrap();
-        assert!(unplayed.is_empty());
-
-        // Add second video, should be unplayed
-        models::upsert_video(&pool, playlist.id, "def456", Some("Song 2"))
-            .await
-            .unwrap();
-        let unplayed = models::get_unplayed_video_ids(&pool, playlist.id)
-            .await
-            .unwrap();
-        assert_eq!(unplayed.len(), 1);
-        assert_eq!(unplayed[0], "def456");
-    }
-
-    #[tokio::test]
-    async fn get_video_file_path_returns_none_for_unnormalized() {
-        let pool = setup().await;
-
-        sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES ('P', 'url')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id) VALUES (1, 'vid1')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Not normalized → should return None.
-        let path = models::get_video_file_path(&pool, 1).await.unwrap();
-        assert!(path.is_none());
-    }
-
-    #[tokio::test]
-    async fn get_video_file_path_returns_path_for_normalized() {
-        let pool = setup().await;
-
-        sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES ('P', 'url')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id, normalized, file_path) VALUES (1, 'vid1', 1, '/cache/song.mp4')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        let path = models::get_video_file_path(&pool, 1).await.unwrap();
-        assert_eq!(path, Some("/cache/song.mp4".to_string()));
-    }
-
-    #[tokio::test]
-    async fn get_video_metadata_returns_song_and_artist() {
-        let pool = setup().await;
-
-        sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES ('P', 'url')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id, song, artist) VALUES (1, 'vid1', 'My Song', 'Artist Name')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        let meta = models::get_video_metadata(&pool, 1).await.unwrap();
-        assert_eq!(
-            meta,
-            Some(("My Song".to_string(), "Artist Name".to_string()))
-        );
-    }
-
-    #[tokio::test]
-    async fn migration_v5_adds_lyrics_columns() {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-        // Verify videos.has_lyrics column exists
-        let row = sqlx::query_scalar::<_, i64>("SELECT has_lyrics FROM videos LIMIT 0")
-            .fetch_optional(&pool)
-            .await;
-        assert!(row.is_ok());
-        // Verify playlists.karaoke_enabled column exists
-        let row = sqlx::query_scalar::<_, i64>("SELECT karaoke_enabled FROM playlists LIMIT 0")
-            .fetch_optional(&pool)
-            .await;
-        assert!(row.is_ok());
-    }
-
-    #[tokio::test]
-    async fn get_active_playlists_includes_ndi_name() {
-        let pool = setup().await;
-
-        sqlx::query("INSERT INTO playlists (name, youtube_url, ndi_output_name) VALUES ('P', 'url', 'SP-test')")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        let playlists = models::get_active_playlists(&pool).await.unwrap();
-        // V13 seeds ytlive, so we now have 2 active playlists
-        assert_eq!(playlists.len(), 2);
-        // Find the test playlist we inserted
-        let p = playlists
-            .iter()
-            .find(|pl| pl.name == "P")
-            .expect("should find test playlist");
-        assert_eq!(p.ndi_output_name, "SP-test");
-    }
-
-    #[tokio::test]
-    async fn migration_v8_downgrades_qwen3_rows_to_lrclib() {
-        let pool = create_memory_pool().await.expect("memory pool");
-        run_migrations(&pool).await.expect("migrations");
-
-        sqlx::query(
-            "INSERT INTO playlists (id, name, youtube_url, ndi_output_name, is_active) \
-             VALUES (99, 'test', 'u', 'n', 1)",
-        )
-        .execute(&pool)
-        .await
-        .unwrap();
-
-        sqlx::query(
-            "INSERT INTO videos (id, playlist_id, youtube_id, title, has_lyrics, lyrics_source) \
-             VALUES (1, 99, 'aaaaaaaaaaa', 't', 1, 'lrclib+qwen3'), \
-                    (2, 99, 'bbbbbbbbbbb', 't', 1, 'lrclib'), \
-                    (3, 99, 'ccccccccccc', 't', 0, NULL)",
-        )
-        .execute(&pool)
-        .await
-        .unwrap();
-
-        // Simulate a row that would have been at schema_version=7 before V8.
-        // run_migrations is idempotent: since the pool already ran through V8,
-        // we have to rewind schema_version and apply V8 manually to prove V8
-        // itself does the right thing on top of V7 state.
-        sqlx::query("UPDATE videos SET lyrics_source = 'lrclib+qwen3' WHERE id = 1")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Re-run migrations: already at version 8 so this is a no-op, but the
-        // point of the assertion below is that after the full migration chain
-        // has been applied, any row currently labeled lrclib+qwen3 was (or
-        // would have been) reset by V8. For an already-migrated DB we instead
-        // apply the V8 SQL directly so the test is deterministic.
-        let v8_sql = MIGRATIONS
-            .iter()
-            .find(|(v, _)| *v == 8)
-            .expect("V8 migration is registered")
-            .1;
-        for stmt in v8_sql.split(';').map(str::trim).filter(|s| !s.is_empty()) {
-            sqlx::query(stmt).execute(&pool).await.unwrap();
-        }
-
-        let (v1_src, v1_has): (Option<String>, i64) =
-            sqlx::query_as("SELECT lyrics_source, has_lyrics FROM videos WHERE id = 1")
-                .fetch_one(&pool)
-                .await
-                .unwrap();
-        assert_eq!(
-            v1_src.as_deref(),
-            Some("lrclib"),
-            "V8 must downgrade the retired combined source value to lrclib"
-        );
-        assert_eq!(
-            v1_has, 1,
-            "has_lyrics must stay 1 after V8 so lyric JSON files are preserved"
-        );
-
-        let (v2_src, v2_has): (Option<String>, i64) =
-            sqlx::query_as("SELECT lyrics_source, has_lyrics FROM videos WHERE id = 2")
-                .fetch_one(&pool)
-                .await
-                .unwrap();
-        assert_eq!(v2_src.as_deref(), Some("lrclib"));
-        assert_eq!(v2_has, 1);
-
-        let (v3_src, v3_has): (Option<String>, i64) =
-            sqlx::query_as("SELECT lyrics_source, has_lyrics FROM videos WHERE id = 3")
-                .fetch_one(&pool)
-                .await
-                .unwrap();
-        assert_eq!(v3_src, None);
-        assert_eq!(v3_has, 0);
-    }
-
-    #[tokio::test]
-    async fn migration_v9_resets_has_lyrics_and_lyrics_source_for_all_rows() {
-        // Seed a DB at V8 with various lyrics_source values, then re-run
-        // migrations to V9 and confirm all rows are back at (0, NULL).
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        sqlx::query(
-            "INSERT INTO playlists (name, youtube_url, ndi_output_name) VALUES ('p', 'u', 'n')",
-        )
-        .execute(&pool)
-        .await
-        .unwrap();
-
-        // `retired_value` is the now-retired combined source literal. Built
-        // at runtime via `concat!` so the unbroken form never appears in
-        // this source file — the CI deletion audit greps for it.
-        let retired_value = concat!("lrclib", "+qwen3");
-        for (yt, src, has) in [
-            ("a1", Some("lrclib"), 1),
-            ("a2", Some("yt_subs+qwen3"), 1),
-            ("a3", Some(retired_value), 1),
-            ("a4", None::<&str>, 0),
-        ] {
-            sqlx::query(
-                "INSERT INTO videos (playlist_id, youtube_id, title, has_lyrics, lyrics_source) \
-                 VALUES (1, ?, 't', ?, ?)",
-            )
-            .bind(yt)
-            .bind(has)
-            .bind(src)
-            .execute(&pool)
-            .await
-            .unwrap();
-        }
-
-        // Rewind schema_version to force V9 to re-run.
-        sqlx::query("DELETE FROM schema_version WHERE version = 9")
-            .execute(&pool)
-            .await
-            .unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        let rows = sqlx::query("SELECT has_lyrics, lyrics_source FROM videos ORDER BY id")
-            .fetch_all(&pool)
-            .await
-            .unwrap();
-        assert_eq!(rows.len(), 4);
-        for row in rows {
-            let hl: i64 = row.get("has_lyrics");
-            let src: Option<String> = row.get("lyrics_source");
-            assert_eq!(hl, 0, "has_lyrics must be 0 after V9");
-            assert_eq!(src, None, "lyrics_source must be NULL after V9");
-        }
-    }
-
-    /// V10 must only reset rows whose `lyrics_source` is `'yt_subs'` (the
-    /// partial half-done state from a deploy where bootstrap failed and
-    /// chunked alignment never ran). LRCLIB and `yt_subs+qwen3` rows
-    /// MUST be left untouched — they're correct as-is and shouldn't pay
-    /// another reprocessing cycle.
-    #[tokio::test]
-    async fn migration_v10_resets_only_partial_yt_subs_rows() {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        sqlx::query(
-            "INSERT INTO playlists (name, youtube_url, ndi_output_name) VALUES ('p', 'u', 'n')",
-        )
-        .execute(&pool)
-        .await
-        .unwrap();
-
-        // Seed AFTER V9 ran: simulate the post-V9 worker writing rows.
-        for (yt, src, has) in [
-            ("a1", Some("lrclib"), 1),        // line-level fallback — keep
-            ("a2", Some("yt_subs+qwen3"), 1), // happy path — keep
-            ("a3", Some("yt_subs"), 1),       // partial — must be reset
-            ("a4", None::<&str>, 0),          // unprocessed — keep at (0, NULL)
-        ] {
-            sqlx::query(
-                "INSERT INTO videos (playlist_id, youtube_id, title, has_lyrics, lyrics_source) \
-                 VALUES (1, ?, 't', ?, ?)",
-            )
-            .bind(yt)
-            .bind(has)
-            .bind(src)
-            .execute(&pool)
-            .await
-            .unwrap();
-        }
-
-        // Force V10 to re-run.
-        sqlx::query("DELETE FROM schema_version WHERE version = 10")
-            .execute(&pool)
-            .await
-            .unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        let rows = sqlx::query(
-            "SELECT youtube_id, has_lyrics, lyrics_source FROM videos ORDER BY youtube_id",
-        )
-        .fetch_all(&pool)
-        .await
-        .unwrap();
-        assert_eq!(rows.len(), 4);
-        let by_id: std::collections::HashMap<String, (i64, Option<String>)> = rows
-            .into_iter()
-            .map(|r| {
-                let yt: String = r.get("youtube_id");
-                (yt, (r.get("has_lyrics"), r.get("lyrics_source")))
-            })
-            .collect();
-
-        assert_eq!(by_id["a1"], (1, Some("lrclib".into())), "lrclib untouched");
-        assert_eq!(by_id["a3"], (0, None), "partial yt_subs reset");
-        assert_eq!(by_id["a4"], (0, None), "already-empty unchanged");
-        // a2 had 'yt_subs+qwen3' which V10 does NOT touch. But V11 DOES
-        // reset it. The test seeds happen AFTER V11 already ran (because
-        // the initial run_migrations got to V11), and only V10 is rewound,
-        // so V11 does NOT re-run and a2 stays untouched here.
-        assert_eq!(
-            by_id["a2"],
-            (1, Some("yt_subs+qwen3".into())),
-            "V10 must not touch yt_subs+qwen3"
-        );
-    }
-
-    /// V11 resets rows whose lyrics_source == 'yt_subs+qwen3' — the
-    /// pre-long-line-split state — so they re-run through the new
-    /// chunking that splits lines with >10 words into sub-chunks.
-    /// LRCLIB rows and partial-reset NULL rows must be left alone.
-    #[tokio::test]
-    async fn migration_v11_resets_only_yt_subs_qwen3_rows() {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        sqlx::query(
-            "INSERT INTO playlists (name, youtube_url, ndi_output_name) VALUES ('p', 'u', 'n')",
-        )
-        .execute(&pool)
-        .await
-        .unwrap();
-
-        for (yt, src, has) in [
-            ("a1", Some("lrclib"), 1),
-            ("a2", Some("yt_subs+qwen3"), 1),
-            ("a3", Some("yt_subs"), 1),
-            ("a4", None::<&str>, 0),
-        ] {
-            sqlx::query(
-                "INSERT INTO videos (playlist_id, youtube_id, title, has_lyrics, lyrics_source) \
-                 VALUES (1, ?, 't', ?, ?)",
-            )
-            .bind(yt)
-            .bind(has)
-            .bind(src)
-            .execute(&pool)
-            .await
-            .unwrap();
-        }
-
-        // Force V11 to re-run.
-        sqlx::query("DELETE FROM schema_version WHERE version = 11")
-            .execute(&pool)
-            .await
-            .unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        let rows = sqlx::query(
-            "SELECT youtube_id, has_lyrics, lyrics_source FROM videos ORDER BY youtube_id",
-        )
-        .fetch_all(&pool)
-        .await
-        .unwrap();
-        let by_id: std::collections::HashMap<String, (i64, Option<String>)> = rows
-            .into_iter()
-            .map(|r| {
-                let yt: String = r.get("youtube_id");
-                (yt, (r.get("has_lyrics"), r.get("lyrics_source")))
-            })
-            .collect();
-
-        assert_eq!(by_id["a1"], (1, Some("lrclib".into())), "lrclib untouched");
-        assert_eq!(
-            by_id["a2"],
-            (0, None),
-            "yt_subs+qwen3 must be reset for re-alignment"
-        );
-        assert_eq!(
-            by_id["a3"],
-            (1, Some("yt_subs".into())),
-            "V11 must not touch partial yt_subs rows"
-        );
-        assert_eq!(by_id["a4"], (0, None), "already-empty unchanged");
-    }
-
-    #[tokio::test]
-    async fn migration_v12_adds_pipeline_version_quality_and_priority() {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-
-        let cols: Vec<String> = sqlx::query("PRAGMA table_info(videos)")
-            .fetch_all(&pool)
-            .await
-            .unwrap()
-            .iter()
-            .map(|r| r.get::<String, _>("name"))
-            .collect();
-
-        assert!(
-            cols.contains(&"lyrics_pipeline_version".to_string()),
-            "missing lyrics_pipeline_version, got: {cols:?}"
-        );
-        assert!(
-            cols.contains(&"lyrics_quality_score".to_string()),
-            "missing lyrics_quality_score, got: {cols:?}"
-        );
-        assert!(
-            cols.contains(&"lyrics_manual_priority".to_string()),
-            "missing lyrics_manual_priority, got: {cols:?}"
-        );
-
-        // Defaults check
-        sqlx::query("INSERT INTO playlists (name, youtube_url) VALUES ('p', 'u')")
-            .execute(&pool)
-            .await
-            .unwrap();
-        sqlx::query("INSERT INTO videos (playlist_id, youtube_id) VALUES (1, 'abc')")
-            .execute(&pool)
-            .await
-            .unwrap();
-        let row = sqlx::query(
-            "SELECT lyrics_pipeline_version, lyrics_manual_priority, lyrics_quality_score \
-             FROM videos WHERE id = 1",
-        )
-        .fetch_one(&pool)
-        .await
-        .unwrap();
-        let pv: i64 = row.get("lyrics_pipeline_version");
-        let mp: i64 = row.get("lyrics_manual_priority");
-        let qs: Option<f64> = row.get("lyrics_quality_score");
-        assert_eq!(pv, 0, "pipeline_version defaults to 0");
-        assert_eq!(mp, 0, "manual_priority defaults to 0");
-        assert_eq!(qs, None, "quality_score defaults to NULL");
-    }
-
-    #[tokio::test]
-    async fn schema_version_reaches_12() {
-        let pool = create_memory_pool().await.unwrap();
-        run_migrations(&pool).await.unwrap();
-        let ver = current_schema_version(&pool).await.unwrap();
-        assert_eq!(ver, 13);
-    }
-
-    #[tokio::test]
-    async fn migration_v13_adds_kind_and_current_position_columns() {
-        let pool = setup().await;
-        let cols: Vec<String> = sqlx::query("PRAGMA table_info(playlists)")
-            .fetch_all(&pool)
-            .await
-            .unwrap()
-            .iter()
-            .map(|r| r.get::<String, _>("name"))
-            .collect();
-        assert!(cols.contains(&"kind".to_string()), "columns: {cols:?}");
-        assert!(
-            cols.contains(&"current_position".to_string()),
-            "columns: {cols:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn migration_v13_creates_playlist_items_table() {
-        let pool = setup().await;
-        let row = sqlx::query(
-            "SELECT name FROM sqlite_master WHERE type='table' AND name='playlist_items'",
-        )
-        .fetch_optional(&pool)
-        .await
-        .unwrap();
-        assert!(row.is_some(), "playlist_items table should exist");
-    }
-
-    #[tokio::test]
-    async fn migration_v13_seeds_ytlive_custom_playlist() {
-        let pool = setup().await;
-        let row = sqlx::query(
-            "SELECT kind, ndi_output_name, playback_mode, is_active, current_position
-             FROM playlists WHERE name = 'ytlive'",
-        )
-        .fetch_one(&pool)
-        .await
-        .unwrap();
-        let kind: String = row.get("kind");
-        let ndi: String = row.get("ndi_output_name");
-        let mode: String = row.get("playback_mode");
-        let is_active: i64 = row.get("is_active");
-        let pos: i64 = row.get("current_position");
-        assert_eq!(kind, "custom");
-        assert_eq!(ndi, "SP-live");
-        assert_eq!(mode, "continuous");
-        assert_eq!(is_active, 1);
-        assert_eq!(pos, 0);
-    }
-}
+mod tests;
